Added SearchForm.jsx component, Bootstrap
