+++ mart-demo/src/OilSearch.jsx
@@ -1,6 +1,5 @@
 import { Link, useParams } from 'react-router-dom';
 import React, { useState, useEffect } from 'react';
-import PropTypes from 'prop-types';
 import { RenderWatercolorHighlights, parseData } from './WatercolorHighlights';
