+++ mart-demo/src/WatercolorHighlights.jsx
@@ -1,7 +1,7 @@
-import { useNavigate, Link } from 'react-router-dom';
 import React, { useState, useEffect } from 'react';
 import dict from './artitems.json';
 import PropTypes from 'prop-types';
+import SearchForm from './SearchForm';
 
 
 
@@ -22,10 +22,6 @@ export function parseData(data, setArtItem) {
 
 export default function WatercolorHighlights() {
     const [artItem, setArtItem] = useState(undefined);
-    const [keyword, setKeyword] = useState(undefined);
-    const [artistOrCulture, setArtistOrCulture] = useState(undefined);
-    const [artTitle, setArtTitle] = useState(undefined);
-    const navigate = useNavigate();
 
     // get random word from json
     let randomInt = getRandomInt(dict.words.length)
@@ -40,41 +36,13 @@ export default function WatercolorHighlights() {
     }, [])
 
 
-    const handleSubmit = () => {
-        if (keyword || artTitle || artistOrCulture) {
-            const params = {
-                "keyword": keyword,
-                "artistOrCulture": artistOrCulture,
-                "artTitle": artTitle
-            }
-            const paramString = JSON.stringify(params)
-            navigate(`/watercolors/search/${paramString}`)
-        }
-
-    }
 
     return (
         <>
             <div>
                 <h1>Watercolor inspirations from the MET collections</h1>
                 <RenderWatercolorHighlights artProfile={artItem} />
-                <p></p>
-                <h3>Search for more</h3>
-                <form>
-                    <label for="key-word"></label>
-                    <input type="text" id="key-word" placeholder='key-word' value={keyword} onChange={e => setKeyword(e.target.value)} />
-                    <p></p>
-                    <label for="art-title"></label>
-                    <input type="text" id="art-title" placeholder='art-title' value={artTitle} onChange={e => setArtTitle(e.target.value)} />
-                    <p></p>
-                    <label for="artistOrCulture"></label>
-                    <input type="text" id="artistOrCulture" placeholder='artistOrCulture' value={artistOrCulture} onChange={e => setArtistOrCulture(e.target.value)} />
-                    <p></p>
-                    <button className="search-button" type="button" onClick={handleSubmit}>Submit</button>
-                    <p></p>
-                </form>
-                <span id='back-link'>
-                    <Link to="/">Homepage</Link></span>
+                <SearchForm medium="watercolors" />
             </div>
         </>
     );
