+++ mart-demo/src/index.js
@@ -4,6 +4,10 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from 'react-router-dom';
+// Bootstrap CSS
+import "bootstrap/dist/css/bootstrap.min.css";
+
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
